Cast port index in label formatting
Update `DmxNode::SetPortLabelDefault` to cast the computed port number to `unsigned` before passing it to `snprintf`. This aligns the argument type with `%u` and avoids format-type warnings.

+++ lib-dmxnode/include/dmxnode.h
@@ -269,7 +269,7 @@ class DmxNode {
         assert(port_index < dmxnode::kMaxPorts);
         auto& port = port_[port_index];
 
-        snprintf(port.label, dmxnode::kPortNameLength - 1, "Port %u", (1U + port_index));
+        snprintf(port.label, dmxnode::kPortNameLength - 1, "Port %u", static_cast<unsigned>((1U + port_index)));
         port.label[dmxnode::kPortNameLength - 1] = '\0';
     }
